+++ api/CHANGELOG.md
@@ -12,6 +12,7 @@
 - Gave keystores its own section in Settings > `Import & Export` and moved relevant settings to that section
 - Added a `Effective memory` pill under the experimental patcher toggle to clarify to the user the max memory the app can use
 - Added more information to the patcher log such as bundle type, and whether the experimental patcher is toggled off or on
+- Added a `Latest` filter and option in the three dot menu to the `Patch bundle discovery`
 
 
 # Bug fixes